Reduce coordinator refactor risk with focused modules
coordinator.rs had accumulated dictation lifecycle, QA panel state, and shared recording resources in one file. This split keeps behavior unchanged while moving the dictation edge/session helpers, QA panel state helpers, and resource ownership helpers into child modules under coordinator/.

Constraint: Issue #296 asks for module boundaries after Rust state-machine coverage is in place

Rejected: Split every helper into separate public modules | larger churn with no behavior benefit

Confidence: high

Scope-risk: moderate

Directive: Keep cross-module wiring in coordinator; leaf modules should remain coordinator-private

Tested: cargo check --manifest-path openless-all/app/src-tauri/Cargo.toml

Tested: cargo test --manifest-path openless-all/app/src-tauri/Cargo.toml --lib

Tested: cargo test --manifest-path openless-all/app/src-tauri/backend-tests/Cargo.toml
